Update Fbl.php
When not if (!empty($this->arfMail['evidence'])) on line 72, then the varialbe $spamMessage Does not gets set so the method getMessageBody cannot be called. eg When mail is nog RFC822 complaint The parse blows up because variable $spamMessage is not set. Another Idea could be to move the setting of the report body up inside the conditional checking the RFC822 complaint But this can have some consequences on the hasRequiredFields;